requirements.txt
Flask
flask-restful
flask-socketio
uwsgi
eventlet
RPi.GPIO
screen
numpy
future
pymavlink
mavproxy
dronekit
